% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tema_ggplot.R
\docType{data}
\name{paleta_metadocencia}
\alias{paleta_metadocencia}
\title{Paleta de colores institucional de MetaDocencia}
\format{
Vector de caracteres con códigos hex.
}
\usage{
paleta_metadocencia
}
\description{
Vector con los cinco colores institucionales en orden recomendado
para uso en gráficos categóricos.
}
\examples{
paleta_metadocencia
scales::show_col(paleta_metadocencia)

}
\keyword{datasets}
