% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tema_ggplot.R
\name{scale_color_metadocencia}
\alias{scale_color_metadocencia}
\alias{scale_colour_metadocencia}
\title{Escala de color discreta con paleta MetaDocencia}
\usage{
scale_color_metadocencia(...)

scale_colour_metadocencia(...)
}
\arguments{
\item{...}{Argumentos adicionales pasados a \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\description{
Para usar con \verb{ggplot2::scale_color_*} y \verb{scale_fill_*}.
}
\examples{
\dontrun{
library(ggplot2)

ggplot(mpg, aes(displ, hwy, color = class)) +
  geom_point(size = 3) +
  scale_color_metadocencia() +
  theme_metadocencia()
}

}
